requirements.txt
pycryptodome
Flask
peewee
netmiko
flask-wtf
pbkdf2
psycopg2-binary
psutil